+++ app/components/AdminPanelV4/tabs/MomentsTab.tsx
@@ -230,6 +230,7 @@ function MomentsTabInner({
           })}
         </div>
       )}
+
     </div>
   );
 }